TypicalTargets: Rename from `TypicalIndexes`
* Add extra targets
